+++ cfdm/core/__init__.py
@@ -11,9 +11,9 @@ datasets and the inspection of CF data model constructs.
 
 """
 
-__date__ = "2025-08-18"
+__date__ = "2025-10-15"
 __cf_version__ = "1.12"
-__version__ = "1.12.3.0"
+__version__ = "1.12.3.1"
 
 from platform import python_version
 
@@ -35,7 +35,7 @@ else:
         )
 
 # Check the version of python
-_minimum_vn = "3.9.0"
+_minimum_vn = "3.10.0"
 if Version(python_version()) < Version(_minimum_vn):
     raise ValueError(
         f"Bad python version: cfdm.core requires python>={_minimum_vn}. "